Refactor pathway analysis instructions and enrichKEGG usage
Removed instructions for starting a Docker session and exiting it since the exercise can run directly on R without docker as the packages have been installed without any errors. 

Updated the section on using the enrichKEGG function for pathway visualization; instead of downloading the KEGG database and running the function locally, the online database is now used directly.